+++ eslint.config.js
@@ -18,6 +18,6 @@ export default tseslint.config(
     },
   },
   {
-    ignores: ["**/dist/**", "**/node_modules/**", "**/.turbo/**"],
+    ignores: ["**/dist/**", "**/node_modules/**", "**/.turbo/**", "**/vitest.config.ts"],
   }
 );